[refactor] 更改module資料夾結構-以section為單位
