+++ src/modules/ExploreCarousel.tsx
