--- src/modules/Section.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function Section() {
-  return <div>Section</div>
-}
